chore: extract trades/alerts/audit from dashboard index (move-only)
