+++ src/routes/dashboard/index.ts
@@ -102,9 +102,12 @@ import { WebullAuth } from '../../infrastructure/webull/WebullAuth'
 import { WebullTokenClient } from '../../infrastructure/webull/WebullTokenClient'
 import { WebullTokenStateClient } from '../../trading/state/WebullTokenStateClient'
 import type { WebullTokenState } from '../../trading/state/WebullTokenStateDO'
-import { JST_FORMATTER, LOG_COPY_ALL_BTN, clampLimit, currencyOfSymbol, displaySymbol, esc, fmtJst, fmtNumber, fmtPct, fmtPctSigned, fmtPriceCcy, formatCooldown, inactiveTooltip, isSymbolInactive, jsonPretty, logCopyRowBtn, messageOf, parseCursor, pillStyle, renderLogCopyScript, renderPaginationNav, safeJsonScript, unavailable } from './shared'
+import { JST_FORMATTER, clampLimit, currencyOfSymbol, displaySymbol, esc, fmtJst, fmtNumber, fmtPct, fmtPctSigned, fmtPriceCcy, formatCooldown, inactiveTooltip, isSymbolInactive, jsonPretty, messageOf, parseCursor, safeJsonScript, unavailable } from './shared'
 import { type DashboardBindings, layout, loadKillSwitchState, renderLayout } from './layout'
+import { tradesBody } from './trades'
 import { cronBody, cronDecisionJson, formatRealizedPnl, loadDecisionRows, renderDecisionLadder, renderDecisionTable } from './cron'
+import { alertsBody, clampAlertLimit, parseAlertsQuery, parseEventTypeFilter, parseSeverityFilter } from './alerts'
+import { auditBody, clampAuditLimit, parseAuditDateFilter, trimQuery } from './audit'
 import { type ChartsBodyArgs, type ChartsBodyGrid, type ChartsBodySymbol, ECHARTS_CDN, type StrategyParamsSnapshot, type SymbolPolicySummary, computeZoomRange, parseChartsTab, parseIsoTimestamp, renderChartsSubnav, renderZoomPresetButtons } from './charts/shared'
 import { MAX_CHART_DECISIONS, type SymbolChartData, type SymbolChartPoint, type SymbolChartRules, densifyHorizontalLine, densifyTrendLine, loadAllSymbolCharts, loadSymbolChart, pickDefaultSymbol } from './charts/loaders'
 import { loadEquityCurve, renderOverviewTab } from './charts/equity'
@@ -112,6 +115,7 @@ import { computePnlHistogram, computeTradeStats, loadDecisionBreakdown, loadTrad
 export { safeJsonScript } from './shared'
 export { localizeReason, renderChartDecisionTrace } from './cron'
 export type { DecisionRow } from './cron'
+export { renderAlertFilterPills } from './alerts'
 export { DEFAULT_ZOOM_WINDOW_MS, computeZoomRange, parseChartsTab, parseIsoTimestamp, renderZoomPresetButtons } from './charts/shared'
 export type { ChartsBodySymbol, ChartsTab, StrategyParamsSnapshot, SymbolPolicySummary } from './charts/shared'
 export { aggregateDailyCloses, anchorJstMidnight, computeChartWindowDays, computeLinearRegressionLine, computeRollingSma, densifyHorizontalLine, densifyTrendLine, deriveOpenPosition, extractSma50, fetchYahooBarsForChart, loadAllSymbolCharts, loadSymbolChart, mergeYahooAndCronPoints, pickDefaultSymbol, resolveFillSide, selectLatestCronSnapshot } from './charts/loaders'
@@ -2972,157 +2976,6 @@ export function renderLastRolledCell(
   return `<span class="ok">${formatted} <small class="muted">(${esc(elapsedLabel)})</small></span>`
 }
 
-/** trade_journal の lifecycle イベント → 日本語ラベル + 色 (#alerts-trades-ui)。 */
-const TRADE_EVENT_LABELS: Record<string, { ja: string; color: string }> = {
-  decision: { ja: '判定', color: '#86868b' },
-  intent: { ja: '注文作成', color: '#46608a' },
-  pre_submit: { ja: '送信記録', color: '#46608a' },
-  post_submit: { ja: '送信応答', color: '#46608a' },
-  fill: { ja: '約定', color: '#057a55' },
-  exit: { ja: '手仕舞い', color: '#b25000' },
-}
-
-/** broker error_code → 短い日本語。未知コードは code をそのまま出す。 */
-const BROKER_ERROR_LABELS: Record<string, string> = {
-  OAUTH_OPENAPI_TICKER_IS_DENY: '銘柄取扱なし',
-  OAUTH_OPENAPI_SELL_QTY_EXCEED_AVAILABLE_QTY: '売却数量超過',
-  OAUTH_OPENAPI_PARAM_ERR: 'パラメータ不正',
-  INVALID_TOKEN: 'トークン無効',
-}
-
-/** errorMessage から error_code らしき token を抜く (JSON / 平文の両対応)。 */
-function extractBrokerErrorCode(message: string): string | null {
-  const fromJson = message.match(/"error_code"\s*:\s*"([A-Z0-9_]+)"/)
-  if (fromJson) return fromJson[1]!
-  const bare = message.match(/\b([A-Z][A-Z0-9_]{6,})\b/)
-  return bare ? bare[1]! : null
-}
-
-function tradesBody(
-  rows: TradeJournalRow[],
-  limit: number,
-  universe?: SymbolUniverse | null,
-  view: 'all' | 'fills' | 'errors' = 'all',
-  before?: number,
-  hasMore = false,
-  filters: { symbol?: string; clientOrderId?: string } = {},
-): string {
-  // symbol / clientOrderId フィルタを view pill・ページネーションの URL に
-  // 伝搬させる (#nav-links)。pill を切り替えても絞り込みが外れないように。
-  const filterQs =
-    (filters.symbol ? `&symbol=${encodeURIComponent(filters.symbol)}` : '') +
-    (filters.clientOrderId ? `&clientOrderId=${encodeURIComponent(filters.clientOrderId)}` : '')
-  const viewPill = (label: string, v: string, active: boolean): string =>
-    `<a href="/dashboard/trades?view=${v}&limit=${limit}${filterQs}" style="margin-right:6px;padding:3px 12px;border-radius:14px;border:1px solid ${active ? '#1d1d1f' : '#d8d8de'};${active ? 'background:#1d1d1f;color:#fff;' : 'background:#fff;'}font-size:12px;text-decoration:none">${esc(label)}</a>`
-  const filterBanner = filters.clientOrderId
-    ? `<p class="muted" style="font-size:12px;margin:0 0 6px">注文 <code>${esc(filters.clientOrderId)}</code> の履歴のみ表示。<a href="/dashboard/cron?clientOrderId=${encodeURIComponent(filters.clientOrderId)}">判定を見る</a> / <a href="/dashboard/trades">全件へ戻る</a></p>`
-    : filters.symbol
-      ? `<p class="muted" style="font-size:12px;margin:0 0 6px">銘柄 <strong>${esc(displaySymbol(filters.symbol, universe))}</strong> のみ表示。<a href="/dashboard/charts?tab=symbol&symbol=${encodeURIComponent(filters.symbol)}">チャートで見る</a> / <a href="/dashboard/cron?symbol=${encodeURIComponent(filters.symbol)}">判定を見る</a> / <a href="/dashboard/trades">全件へ戻る</a></p>`
-      : ''
-  const pills = `<nav style="margin-bottom:10px;display:flex;align-items:center;flex-wrap:wrap;gap:2px">${viewPill('全イベント', 'all', view === 'all')}${viewPill('約定・手仕舞い', 'fills', view === 'fills')}${viewPill('エラー', 'errors', view === 'errors')}<span class="muted" style="font-size:12px;margin:0 8px">${rows.length} 件 (limit=${limit})</span>${rows.length > 0 ? LOG_COPY_ALL_BTN : ''}</nav>`
-  if (rows.length === 0) {
-    return `${filterBanner}${pills}<p class="muted">該当するレコードがありません。</p>`
-  }
-  const tbody = rows
-    .map((r) => {
-      const ev = TRADE_EVENT_LABELS[r.tradeEventType] ?? { ja: r.tradeEventType, color: '#86868b' }
-      const eventCell = `<span title="${esc(r.tradeEventType)}" style="color:${ev.color};font-weight:600">● ${esc(ev.ja)}</span>`
-      const symbolText = r.symbol ? displaySymbol(r.symbol, universe) : null
-      const inactive = r.symbol ? isSymbolInactive(r.symbol, universe) : false
-      // ▼ は同一銘柄の約定だけに絞り込み、「判定」は clientOrderId でこの注文の
-      // 判定行 (cron) へ飛ぶ逆リンク (#nav-links)。
-      const symbolCell = r.symbol
-        ? `<a href="/dashboard/charts?tab=symbol&symbol=${encodeURIComponent(r.symbol)}"${inactive ? ` title="${esc(inactiveTooltip(r.symbol, universe))}"` : ''} style="text-decoration:none"><strong${inactive ? ' class="symbol-disabled"' : ''}>${esc(symbolText!)}</strong></a> <a href="/dashboard/trades?symbol=${encodeURIComponent(r.symbol)}" class="muted" title="この銘柄の約定だけに絞り込み" style="font-size:11px;text-decoration:none">▼</a>`
-        : '<span class="muted">—</span>'
-      const decisionLink = r.clientOrderId
-        ? ` <a href="/dashboard/cron?clientOrderId=${encodeURIComponent(r.clientOrderId)}" class="muted" title="この注文の判定 (戦略判定ログ) を見る" style="font-size:11px">判定→</a>`
-        : ''
-      const sideCell =
-        r.side === 'BUY'
-          ? `<span class="ok" style="font-weight:700">買</span> <span class="muted" style="font-size:11px">BUY</span>`
-          : r.side === 'SELL'
-            ? `<span class="err" style="font-weight:700">売</span> <span class="muted" style="font-size:11px">SELL</span>`
-            : '<span class="muted">—</span>'
-      // 数量: 発注数量 → 約定数量。一致なら 1 つだけ、部分約定が見えるように。
-      const qtyCell =
-        r.filledQty !== null && r.quantity !== null && r.filledQty !== r.quantity
-          ? `${esc(r.quantity)} → <strong>${esc(r.filledQty)}</strong>`
-          : r.filledQty !== null
-            ? `${esc(r.filledQty)}`
-            : r.quantity !== null
-              ? `${esc(r.quantity)}`
-              : '—'
-      const priceCell =
-        r.filledPrice !== null
-          ? fmtNumber(r.filledPrice, 2)
-          : r.limitPrice !== null
-            ? `<span class="muted" title="指値 (未約定)">指 ${fmtNumber(r.limitPrice, 2)}</span>`
-            : '—'
-      const pnlCell =
-        r.realizedPnl !== null
-          ? `${formatRealizedPnl(r.realizedPnl)}${r.exitReason ? ` <span class="muted" style="font-size:11px">${esc(r.exitReason)}</span>` : ''}`
-          : '<span class="muted">—</span>'
-      // 状態: エラーは短い日本語 + code、全文は <details>。enum はそのまま残す
-      // (broker API と grep で突き合わせる運用のため title / details に保持)。
-      let statusCell: string
-      const errorText = r.errorMessage ?? r.errorClass
-      if (errorText) {
-        const code = extractBrokerErrorCode(errorText)
-        const short = code ? (BROKER_ERROR_LABELS[code] ?? code) : (r.errorClass ?? 'エラー')
-        statusCell = `<span style="${pillStyle('#fdecec', '#c22')}">エラー: ${esc(short)}</span>
-          <details style="margin-top:2px"><summary class="muted" style="font-size:11px;cursor:pointer">全文</summary><code style="font-size:11px;white-space:pre-wrap;word-break:break-all">${esc(errorText)}</code></details>`
-      } else if (r.brokerStatus === 'FILLED') {
-        statusCell = `<span style="${pillStyle('#e6f6ec', '#057a55')}">約定</span>`
-      } else if (r.brokerStatus) {
-        statusCell = `<span style="${pillStyle('#fff4e6', '#b25000')}" title="${esc(r.brokerStatus)}">${esc(r.brokerStatus)}</span>`
-      } else {
-        statusCell = '<span class="muted">—</span>'
-      }
-      const modeCell =
-        r.mode === 'LIVE'
-          ? `<span style="${pillStyle('#fdecec', '#c22')}">実発注</span>`
-          : r.mode === 'DRY_RUN'
-            ? `<span style="${pillStyle('#f3f3f5', '#86868b')}">DRY</span>`
-            : '<span class="muted">—</span>'
-      return `<tr style="font-size:13px">
-        <td>${logCopyRowBtn(r.id)}</td>
-        <td class="muted" style="white-space:nowrap">${esc(fmtJst(r.timestamp))}</td>
-        <td style="white-space:nowrap">${eventCell}${decisionLink}</td>
-        <td>${symbolCell}</td>
-        <td style="white-space:nowrap">${sideCell}</td>
-        <td style="text-align:right" class="bp-num">${qtyCell}</td>
-        <td style="text-align:right" class="bp-num">${priceCell}</td>
-        <td style="text-align:right" class="bp-num">${pnlCell}</td>
-        <td>${statusCell}</td>
-        <td>${modeCell}</td>
-      </tr>`
-    })
-    .join('')
-  return `${filterBanner}${pills}
-  <table>
-    <thead><tr style="font-size:12px">
-      <th></th><th>日時 (JST)</th><th>イベント</th><th>銘柄</th><th>売買</th>
-      <th style="text-align:right">数量</th><th style="text-align:right">単価</th><th style="text-align:right">実現損益</th><th>状態</th><th>モード</th>
-    </tr></thead>
-    <tbody>${tbody}</tbody>
-  </table>
-  ${renderPaginationNav({
-    baseHref: `/dashboard/trades?view=${view}&limit=${limit}${filterQs}`,
-    before,
-    lastId: rows.length > 0 ? rows[rows.length - 1]!.id : undefined,
-    hasMore,
-  })}
-  ${safeJsonScript('__tradesCopy', {
-    meta: {
-      page: 'trade_journal (約定履歴)',
-      filter: `view=${view}, limit=${limit}${filters.symbol ? `, symbol=${filters.symbol}` : ''}${filters.clientOrderId ? `, clientOrderId=${filters.clientOrderId}` : ''}`,
-      generatedAt: new Date().toISOString(),
-    },
-    rows,
-  })}
-  ${renderLogCopyScript('__tradesCopy')}`
-}
-
 function configBody(
   global: Awaited<ReturnType<typeof loadGlobalConfigFrom>>,
   universe: Awaited<ReturnType<typeof loadSymbolUniverse>>,
@@ -3368,332 +3221,6 @@ function formatConfigValue(v: unknown): string {
   if (v === null || v === undefined) return '—'
   if (typeof v === 'boolean') return v ? 'true' : 'false'
   return String(v)
-}
-
-/**
- * `?limit=N` を 1〜500 の範囲に丸める。`/dashboard/alerts` 専用 (cron 系の
- * `clampLimit` は既定 50 / max 200 で別ロール)。
- */
-function clampAlertLimit(raw: string | undefined): number {
-  const n = raw === undefined ? 100 : Number.parseInt(raw, 10)
-  if (!Number.isFinite(n) || n <= 0) return 100
-  return Math.min(n, 500)
-}
-
-/**
- * `c.req.url` から `URLSearchParams` を取り出す。filter pill が他の query
- * (例: `limit=500`) を保持するために使う (CodeRabbit #210)。
- *
- * URL 構築失敗時は空 URLSearchParams にフォールバック。
- */
-function parseAlertsQuery(rawUrl: string): URLSearchParams {
-  try {
-    return new URL(rawUrl).searchParams
-  } catch {
-    return new URLSearchParams()
-  }
-}
-
-const SEVERITY_VALUES: ReadonlyArray<NotificationSeverity> = ['critical', 'warning', 'info']
-
-function parseSeverityFilter(raw: string | undefined): NotificationSeverity[] {
-  if (!raw) return []
-  return raw
-    .split(',')
-    .map((s) => s.trim().toLowerCase())
-    .filter((s): s is NotificationSeverity =>
-      (SEVERITY_VALUES as readonly string[]).includes(s),
-    )
-}
-
-const EVENT_TYPE_VALUES: ReadonlyArray<NotificationEvent['type']> = [
-  'TRADE',
-  'ERROR',
-  'STATE_CHANGE',
-]
-
-function parseEventTypeFilter(raw: string | undefined): NotificationEvent['type'] | undefined {
-  if (!raw) return undefined
-  const upper = raw.trim().toUpperCase() as NotificationEvent['type']
-  return (EVENT_TYPE_VALUES as readonly string[]).includes(upper) ? upper : undefined
-}
-
-interface AlertsBodyArgs {
-  rows: AlertRow[]
-  limit: number
-  severityFilter: NotificationSeverity[]
-  eventTypeFilter: NotificationEvent['type'] | undefined
-  /** 現在の query string。filter pill が他の param (limit 等) を保持するために使う。 */
-  currentQuery: URLSearchParams
-  /** symbol 列の表示を JP 銘柄向け 番号-会社名 形式にするための universe (load 失敗は null)。 */
-  universe?: SymbolUniverse | null
-  before?: number
-  hasMore?: boolean
-}
-
-/**
- * `/dashboard/alerts` の HTML 本文 (#141)。
- *
- *   - severity ピル (critical / warning / info / 全件) で絞り込み
- *   - event type ピル (TRADE / ERROR / STATE_CHANGE / 全件) で絞り込み
- *   - 表示は最新 100 件 (`?limit=N` で 1〜500)
- *   - 行クリックで Slack/Discord に出したのと同じ message を JST 時刻と一緒に確認
- */
-/** severity → 日本語 pill (#alerts-trades-ui)。 */
-const ALERT_SEVERITY_PILLS: Record<string, { ja: string; bg: string; fg: string }> = {
-  critical: { ja: '重大', bg: '#fdecec', fg: '#c22' },
-  warning: { ja: '警告', bg: '#fff4e6', fg: '#b25000' },
-  info: { ja: '情報', bg: '#eef2f8', fg: '#46608a' },
-}
-
-/** event type → 日本語。 */
-const ALERT_EVENT_LABELS: Record<string, string> = {
-  ERROR: 'エラー',
-  TRADE: '売買',
-  STATE_CHANGE: '設定変更',
-}
-
-/** 長い message は先頭を出して残りを <details> に畳む閾値。 */
-const ALERT_MESSAGE_FOLD = 160
-
-function alertsBody(args: AlertsBodyArgs): string {
-  const { rows, limit, severityFilter, eventTypeFilter, currentQuery, universe, before, hasMore = false } = args
-  const filterPills = renderAlertFilterPills(severityFilter, eventTypeFilter, currentQuery)
-  const countLine = `<span class="muted" style="font-size:12px;margin-right:8px">${rows.length} 件 (limit=${limit}, max 500)</span>${rows.length > 0 ? LOG_COPY_ALL_BTN : ''}`
-  if (rows.length === 0) {
-    return `${filterPills}${countLine}<p class="muted">該当するアラートはありません。</p>`
-  }
-  const tbody = rows
-    .map((r) => {
-      const sev = ALERT_SEVERITY_PILLS[r.severity] ?? { ja: r.severity, bg: '#f3f3f5', fg: '#86868b' }
-      const sevCell = `<span title="${esc(r.severity)}" style="${pillStyle(sev.bg, sev.fg)}">${esc(sev.ja)}</span>`
-      const eventCell = `<span title="${esc(r.eventType)}" style="font-size:12px">${esc(ALERT_EVENT_LABELS[r.eventType] ?? r.eventType)}</span>`
-      const symbolInactive = r.symbol ? isSymbolInactive(r.symbol, universe) : false
-      const symbolCell = r.symbol
-        ? `<a href="/dashboard/charts?tab=symbol&symbol=${encodeURIComponent(r.symbol)}"${symbolInactive ? ` title="${esc(inactiveTooltip(r.symbol, universe))}"` : ''} style="text-decoration:none"><strong${symbolInactive ? ' class="symbol-disabled"' : ''}>${esc(displaySymbol(r.symbol, universe))}</strong></a>`
-        : '<span class="muted">—</span>'
-      // broker エラーは error_code を短い日本語にして先頭へ。message 全文は
-      // 長ければ畳む (enum / 原文は grep 突き合わせ用に details に保持)。
-      const code = r.eventType === 'ERROR' ? extractBrokerErrorCode(r.message) : null
-      const shortLabel = code ? (BROKER_ERROR_LABELS[code] ?? code) : null
-      const messageBody =
-        r.message.length > ALERT_MESSAGE_FOLD
-          ? `${esc(r.message.slice(0, ALERT_MESSAGE_FOLD))}…<details style="margin-top:2px"><summary class="muted" style="font-size:11px;cursor:pointer">全文</summary><code style="font-size:11px;white-space:pre-wrap;word-break:break-all">${esc(r.message)}</code></details>`
-          : esc(r.message)
-      // <details> (block) を含み得るので外側は div (CodeRabbit #469)。
-      const messageCell = `${shortLabel ? `<span style="${pillStyle('#fdecec', '#c22')}">${esc(shortLabel)}</span>` : ''}<div style="font-size:12px">${messageBody}</div>`
-      const causeCell = r.cause
-        ? `<code style="font-size:11px">${esc(r.cause)}</code>`
-        : '<span class="muted">—</span>'
-      return `<tr style="font-size:13px;vertical-align:top">
-        <td>${logCopyRowBtn(r.id)}</td>
-        <td class="muted" style="white-space:nowrap">${esc(fmtJst(r.timestamp))}</td>
-        <td>${sevCell}</td>
-        <td>${eventCell}</td>
-        <td>${symbolCell}</td>
-        <td>${causeCell}</td>
-        <td>${messageCell}</td>
-        <td class="muted"><code style="font-size:11px">${esc(r.requestId ?? '—')}</code></td>
-      </tr>`
-    })
-    .join('')
-  return `${filterPills}${countLine}
-  <table>
-    <thead><tr style="font-size:12px">
-      <th></th><th>日時 (JST)</th><th>重要度</th><th>種別</th><th>銘柄</th><th>要因</th><th>内容</th><th>requestId</th>
-    </tr></thead>
-    <tbody>${tbody}</tbody>
-  </table>
-  ${renderPaginationNav({
-    baseHref: buildAlertBaseHref(limit, severityFilter, eventTypeFilter),
-    before,
-    lastId: rows.length > 0 ? rows[rows.length - 1]!.id : undefined,
-    hasMore,
-  })}
-  ${safeJsonScript('__alertsCopy', {
-    meta: {
-      page: 'notification_emit_log (アラート)',
-      filter:
-        severityFilter.length === 0 && eventTypeFilter === undefined
-          ? '全件'
-          : `severity=${severityFilter.join(',') || 'all'}, eventType=${eventTypeFilter ?? 'all'}`,
-      generatedAt: new Date().toISOString(),
-    },
-    rows,
-  })}
-  ${renderLogCopyScript('__alertsCopy')}`
-}
-
-function buildAlertBaseHref(
-  limit: number,
-  severityFilter: NotificationSeverity[],
-  eventTypeFilter: NotificationEvent['type'] | undefined,
-): string {
-  const params: string[] = [`limit=${limit}`]
-  if (severityFilter.length > 0) params.push(`severity=${severityFilter.join(',')}`)
-  if (eventTypeFilter) params.push(`eventType=${eventTypeFilter}`)
-  return `/dashboard/alerts?${params.join('&')}`
-}
-
-/**
- * `/dashboard/alerts` の severity / eventType filter ピルを描画する。
- *
- * `currentQuery` の他 param (例: `limit=500`) は preserve したまま、対象 key
- * のみを差し替える / 削除する (CodeRabbit #210)。
- *
- * Exported for unit test (URL preservation).
- */
-export function renderAlertFilterPills(
-  active: NotificationSeverity[],
-  activeEventType: NotificationEvent['type'] | undefined,
-  currentQuery: URLSearchParams,
-): string {
-  const buildHref = (updatedKey: string, updatedValue: string | null): string => {
-    const next = new URLSearchParams(currentQuery)
-    if (updatedValue === null) next.delete(updatedKey)
-    else next.set(updatedKey, updatedValue)
-    const qs = next.toString()
-    return qs.length === 0 ? '/dashboard/alerts' : `/dashboard/alerts?${qs}`
-  }
-  const pill = (label: string, href: string, isActive: boolean): string =>
-    `<a href="${esc(href)}" style="margin-right:6px;${isActive ? 'background:#1d1d1f;color:#fff;' : ''}">${esc(label)}</a>`
-  const sev = [
-    pill('全 severity', buildHref('severity', null), active.length === 0),
-    pill(
-      'critical',
-      buildHref('severity', 'critical'),
-      active.length === 1 && active[0] === 'critical',
-    ),
-    pill(
-      'warning',
-      buildHref('severity', 'warning'),
-      active.length === 1 && active[0] === 'warning',
-    ),
-    pill(
-      'critical+warning',
-      buildHref('severity', 'critical,warning'),
-      active.length === 2 && active.includes('critical') && active.includes('warning'),
-    ),
-    pill('info', buildHref('severity', 'info'), active.length === 1 && active[0] === 'info'),
-  ].join('')
-  const ev = [
-    pill('全 type', buildHref('eventType', null), activeEventType === undefined),
-    pill('ERROR', buildHref('eventType', 'ERROR'), activeEventType === 'ERROR'),
-    pill('TRADE', buildHref('eventType', 'TRADE'), activeEventType === 'TRADE'),
-    pill(
-      'STATE_CHANGE',
-      buildHref('eventType', 'STATE_CHANGE'),
-      activeEventType === 'STATE_CHANGE',
-    ),
-  ].join('')
-  return `<nav style="margin-bottom:12px">${sev}<span class="muted" style="margin:0 8px">|</span>${ev}</nav>`
-}
-
-interface AuditBodyArgs {
-  rows: ConfigAuditRow[]
-  limit: number
-  actorFilter: string | undefined
-  endpointFilter: string | undefined
-  /** Raw query string values for the form inputs (passthrough so a typo round-trips). */
-  fromFilter: string
-  toFilter: string
-  before?: number
-  hasMore?: boolean
-}
-
-/**
- * `/dashboard/audit` の HTML 本文 (#274)。
- *
- *   - 直近 100 件 (`?limit=N` で 1〜500)
- *   - actor / endpoint / from / to で絞り込み (GET form)
- *   - before_json / after_json は `<details>` で展開表示
- */
-function auditBody(args: AuditBodyArgs): string {
-  const { rows, limit, actorFilter, endpointFilter, fromFilter, toFilter, before, hasMore = false } = args
-  const form = `<form method="get" action="/dashboard/audit" style="margin-bottom:12px;display:flex;gap:8px;flex-wrap:wrap;align-items:flex-end">
-  <label>actor<br><input name="actor" value="${esc(actorFilter ?? '')}" placeholder="ai-agent" style="padding:4px 8px"></label>
-  <label>endpoint<br><input name="endpoint" value="${esc(endpointFilter ?? '')}" placeholder="/admin/symbols/:symbol/seed-cash" style="padding:4px 8px;min-width:280px"></label>
-  <label>from<br><input name="from" type="date" value="${esc(fromFilter)}" style="padding:4px 8px"></label>
-  <label>to<br><input name="to" type="date" value="${esc(toFilter)}" style="padding:4px 8px"></label>
-  <label>limit<br><input name="limit" type="number" min="1" max="500" value="${limit}" style="padding:4px 8px;width:90px"></label>
-  <button type="submit" style="padding:6px 14px">絞り込み</button>
-  <a href="/dashboard/audit" style="padding:6px 14px;text-decoration:none">リセット</a>
-</form>`
-  const header = `<p class="muted">直近 ${rows.length} 件 (limit=${limit}, max 500)。状態変更系 admin POST の before/after diff。before == after の no-op 呼び出しは記録されません。</p>`
-  if (rows.length === 0) {
-    return `${header}${form}<p class="muted">該当する監査ログは見つかりませんでした。</p>`
-  }
-  const tbody = rows
-    .map((r) => {
-      return `<tr>
-        <td class="muted">${esc(fmtJst(r.timestamp))}</td>
-        <td><strong>${esc(r.actor)}</strong></td>
-        <td><code>${esc(r.endpoint)}</code></td>
-        <td>${esc(r.targetKey ?? '-')}</td>
-        <td><details><summary class="muted">before</summary><pre style="margin:4px 0 0;white-space:pre-wrap;word-break:break-word;font-size:12px;background:#fafafa;padding:6px;border-radius:4px">${esc(formatAuditJson(r.beforeJson))}</pre></details></td>
-        <td><details><summary class="muted">after</summary><pre style="margin:4px 0 0;white-space:pre-wrap;word-break:break-word;font-size:12px;background:#fafafa;padding:6px;border-radius:4px">${esc(formatAuditJson(r.afterJson))}</pre></details></td>
-        <td class="muted"><code>${esc(r.requestId ?? '-')}</code></td>
-      </tr>`
-    })
-    .join('')
-  const auditBaseParams: string[] = [`limit=${limit}`]
-  if (actorFilter) auditBaseParams.push(`actor=${encodeURIComponent(actorFilter)}`)
-  if (endpointFilter) auditBaseParams.push(`endpoint=${encodeURIComponent(endpointFilter)}`)
-  if (fromFilter) auditBaseParams.push(`from=${encodeURIComponent(fromFilter)}`)
-  if (toFilter) auditBaseParams.push(`to=${encodeURIComponent(toFilter)}`)
-  return `${header}${form}
-  <table>
-    <thead><tr>
-      <th>timestamp (JST)</th><th>actor</th><th>endpoint</th><th>target</th><th>before</th><th>after</th><th>requestId</th>
-    </tr></thead>
-    <tbody>${tbody}</tbody>
-  </table>
-  ${renderPaginationNav({
-    baseHref: `/dashboard/audit?${auditBaseParams.join('&')}`,
-    before,
-    lastId: rows.length > 0 ? rows[rows.length - 1]!.id : undefined,
-    hasMore,
-  })}`
-}
-
-/**
- * `?limit=N` を 1〜500 に丸める。`/dashboard/audit` 既定 100。
- */
-function clampAuditLimit(raw: string | undefined): number {
-  const n = raw === undefined ? 100 : Number.parseInt(raw, 10)
-  if (!Number.isFinite(n) || n <= 0) return 100
-  return Math.min(n, 500)
-}
-
-function trimQuery(raw: string | undefined): string | undefined {
-  if (raw === undefined) return undefined
-  const trimmed = raw.trim()
-  return trimmed.length === 0 ? undefined : trimmed
-}
-
-/**
- * `YYYY-MM-DD` 日付フィルタを ISO timestamp に展開。`isEnd=true` は `T23:59:59.999Z`、
- * false は `T00:00:00.000Z` を付ける (UTC base — 監査ログの timestamp は
- * ISO UTC で書かれる)。文法が合わない値は undefined を返す (フィルタ skip)。
- */
-function parseAuditDateFilter(raw: string | undefined, isEnd: boolean): string | undefined {
-  if (raw === undefined) return undefined
-  const trimmed = raw.trim()
-  if (trimmed === '') return undefined
-  if (!/^\d{4}-\d{2}-\d{2}$/.test(trimmed)) return undefined
-  return isEnd ? `${trimmed}T23:59:59.999Z` : `${trimmed}T00:00:00.000Z`
-}
-
-/**
- * `before_json` / `after_json` を整形して表示。JSON parse が成功すれば 2-space
- * indent、失敗 (= マイグレ前の raw 文字列など) は原文をそのまま返す。
- */
-function formatAuditJson(raw: string): string {
-  try {
-    return JSON.stringify(JSON.parse(raw), null, 2)
-  } catch {
-    return raw
-  }
 }
 
 function chartsBody(args: ChartsBodyArgs): string {
